requirements.txt
# Discord Bot
discord.py>=2.3.0

# Web Server
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.0

# QR Code
qrcode>=7.4.0
pillow>=10.0.0

# Excel Export
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0
